添加测试用例时查重
# 添加测试用例查重
# 启动容器时检查端口占用情况

+++ start-dev-dockers.sh
@@ -1,6 +1,11 @@
 db=content_db
 app=content_app
 port=8000
+if [ -n "$(lsof -i:$port)" ]
+then
+  echo port $port is used, failed
+  exit 1
+fi
 docker rm -f $db
 docker rm -f $app
 ./mkmemdisk.sh
